Refactor #27- move package
