+++ jdbc/src/main/java/hello/jdbc/repository/MemberRepositoryV4_2.java
@@ -1,4 +1,4 @@
-package hello.jdbc.repository.ex;
+package hello.jdbc.repository;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
