+++ jdbc/src/test/java/hello/jdbc/service/MemberServiceV4Test.java
@@ -15,7 +15,7 @@ import org.springframework.context.annotation.Bean;
 
 import hello.jdbc.domain.Member;
 import hello.jdbc.repository.MemberRepository;
-import hello.jdbc.repository.ex.MemberRepositoryV4_2;
+import hello.jdbc.repository.MemberRepositoryV4_2;
 import lombok.extern.slf4j.Slf4j;
 
 /**
